scripts/requirements.txt
black==19.10b0
mistune==2.0.0a1
attrs==19.1.0
PyYAML==5.2
